feat: move functions to collector
